Feat : 마이페이지 소개팅 등록 토글 활성화

+++ Frontend/ssafy-town/src/components/sidebar/8mypage.tsx
@@ -58,6 +58,7 @@ const Mypage: React.FC = () => {
   const dispatch = useDispatch();
   // const isSidebarOpen = useSelector((state: AppState) => state.isSidebarOpen);//사이드바 오픈여부
   const [user, setUser] = useState(userdata)
+  const [sogae, setSogae] = useState(true)
   const isModalOpen = useSelector((state: AppState) => state.isModalOpen);// 모달창 오픈여부 (알림, 로그아웃)
 
   const toggleHandler = () => {
@@ -114,28 +115,45 @@ const Mypage: React.FC = () => {
       })
   }, [])
 
-  // 소개팅 등록한 경우 등록철회
-  const unregistMeeting = () => {
-    console.log("이제 소개팅 안할래!");
+  // 소개팅 등록 상태 전환
+  const onregistMeeting = () => {
     const userIdx = localStorage.getItem('userIdx');
     const userToken = localStorage.getItem('userToken');
-    axios({
-      method: 'delete',
-      url: `https://i9b206.p.ssafy.io:9090/date/release/${userIdx}`,
-      headers: {
-        Authorization: 'Bearer ' + userToken
-      }
-    })
-      .then(res => {
-        console.log(res)
-        alert("소개팅 등록해제 성공");
+    
+    if (sogae) {
+      axios({
+        method: 'delete',
+        url: `https://i9b206.p.ssafy.io:9090/date/release/${userIdx}`,
+        headers: {
+          Authorization: 'Bearer ' + userToken
+        }
       })
-      .catch(err => {
-        console.log(err);
-        alert("소개팅 등록해제 실패");
+        .then(() => {
+          setSogae(false)
+        })
+        .catch(() => {
+          alert("소개팅 등록해제 실패");
+        })
+    }
+    else {
+      axios({
+        method: 'post',
+        url: `https://i9b206.p.ssafy.io:9090/date/register/${userIdx}`,
+        headers: {
+          Authorization: 'Bearer ' + userToken
+        }
       })
+      .then(() => {
+          setSogae(true)
+        })
+        .catch(() => {
+          alert("소개팅 등록 실패");
+        })
+    }
   }
 
+
+  
   return (
     <div>
       <div className='sidebar_modal' id={mypage_css.modal}>
@@ -165,14 +183,13 @@ const Mypage: React.FC = () => {
           </p>
 
           <p className={mypage_css.mypage_toggle}>소개팅 등록
-            <ToggleContainer onClick={()=>{toggleHandler()}}>
-              <div className={`toggle-container ${user.invite === 'true' ? "toggle--checked" : null}`}/>
-              <div className={`${mypage_css.mypage_toggle} toggle-circle ${ user.invite==='true' ? "toggle--checked" : null}`}>
-                 {user.invite === 'true' ? 'On':'Off'}</div>
+            <ToggleContainer onClick={()=>{onregistMeeting()}}>
+              <div className={`toggle-container ${sogae ? "toggle--checked" : null}`}/>
+              <div className={`${mypage_css.mypage_toggle} toggle-circle ${ sogae ? "toggle--checked" : null}`}>
+                 {sogae ? 'On':'Off'}</div>
             </ToggleContainer>
           </p>
 
-          <button className={mypage_css.button} onClick={unregistMeeting}>소개팅 등록 취소</button>
           <button className={mypage_css.button} onClick={() => dispatch(setModal('Re최초설문'))}>최초 설문 수정</button>
         </div>
         {isModalOpen === '회원정보수정3' ? <EditAcount user={user} /> :
